+++ packages/backend-rs/src/model/entity/mod.rs
@@ -51,6 +51,7 @@ pub mod registration_ticket;
 pub mod registry_item;
 pub mod relay;
 pub mod renote_muting;
+pub mod reply_muting;
 pub mod sea_orm_active_enums;
 pub mod signin;
 pub mod sw_subscription;