+++ packages/backend-rs/src/model/entity/prelude.rs
@@ -49,6 +49,7 @@ pub use super::registration_ticket::Entity as RegistrationTicket;
 pub use super::registry_item::Entity as RegistryItem;
 pub use super::relay::Entity as Relay;
 pub use super::renote_muting::Entity as RenoteMuting;
+pub use super::reply_muting::Entity as ReplyMuting;
 pub use super::signin::Entity as Signin;
 pub use super::sw_subscription::Entity as SwSubscription;
 pub use super::used_username::Entity as UsedUsername;