+++ packages/backend/src/db/postgre.ts
@@ -24,6 +24,7 @@ import { Following } from "@/models/entities/following.js";
 import { Instance } from "@/models/entities/instance.js";
 import { Muting } from "@/models/entities/muting.js";
 import { RenoteMuting } from "@/models/entities/renote-muting.js";
+import { ReplyMuting } from "@/models/entities/reply-muting.js";
 import { SwSubscription } from "@/models/entities/sw-subscription.js";
 import { Blocking } from "@/models/entities/blocking.js";
 import { UserList } from "@/models/entities/user-list.js";
@@ -137,6 +138,7 @@ export const entities = [
 	FollowRequest,
 	Muting,
 	RenoteMuting,
+	ReplyMuting,
 	Blocking,
 	Note,
 	NoteEdit,