+++ packages/backend/src/models/index.ts
@@ -26,6 +26,7 @@ import { UserGroupInvitationRepository } from "./repositories/user-group-invitat
 import { FollowRequestRepository } from "./repositories/follow-request.js";
 import { MutingRepository } from "./repositories/muting.js";
 import { RenoteMutingRepository } from "./repositories/renote-muting.js";
+import { ReplyMutingRepository } from "./repositories/reply-muting.js";
 import { BlockingRepository } from "./repositories/blocking.js";
 import { NoteReactionRepository } from "./repositories/note-reaction.js";
 import { NotificationRepository } from "./repositories/notification.js";
@@ -103,6 +104,7 @@ export const Notifications = NotificationRepository;
 export const Metas = db.getRepository(Meta);
 export const Mutings = MutingRepository;
 export const RenoteMutings = RenoteMutingRepository;
+export const ReplyMutings = ReplyMutingRepository;
 export const Blockings = BlockingRepository;
 export const SwSubscriptions = db.getRepository(SwSubscription);
 export const Hashtags = HashtagRepository;