+++ packages/backend/src/server/api/endpoints/notes/global-timeline.ts
@@ -9,6 +9,7 @@ import { generateRepliesQuery } from "@/server/api/common/generate-replies-query
 import { generateMutedNoteQuery } from "@/server/api/common/generate-muted-note-query.js";
 import { generateBlockedUserQuery } from "@/server/api/common/generate-block-query.js";
 import { generateMutedUserRenotesQueryForNotes } from "@/server/api/common/generated-muted-renote-query.js";
+import { generateMutedUserRepliesQueryForNotes } from "@/server/api/common/generated-muted-reply-query.js";
 
 export const meta = {
 	tags: ["notes"],
@@ -93,11 +94,12 @@ export default define(meta, paramDef, async (ps, user) => {
 		.leftJoinAndSelect("renoteUser.banner", "renoteUserBanner");
 
 	generateRepliesQuery(query, ps.withReplies, user);
-	if (user) {
+	if (user != null) {
 		generateMutedUserQuery(query, user);
 		generateMutedNoteQuery(query, user);
 		generateBlockedUserQuery(query, user);
 		generateMutedUserRenotesQueryForNotes(query, user);
+		generateMutedUserRepliesQueryForNotes(query, user);
 	}
 
 	if (ps.withFiles) {