+++ packages/backend/src/server/api/endpoints/notes/local-timeline.ts
@@ -12,6 +12,7 @@ import { generateMutedNoteQuery } from "@/server/api/common/generate-muted-note-
 import { generateChannelQuery } from "@/server/api/common/generate-channel-query.js";
 import { generateBlockedUserQuery } from "@/server/api/common/generate-block-query.js";
 import { generateMutedUserRenotesQueryForNotes } from "@/server/api/common/generated-muted-renote-query.js";
+import { generateMutedUserRepliesQueryForNotes } from "@/server/api/common/generated-muted-reply-query.js";
 
 export const meta = {
 	tags: ["notes"],
@@ -104,10 +105,13 @@ export default define(meta, paramDef, async (ps, user) => {
 	generateChannelQuery(query, user);
 	generateRepliesQuery(query, ps.withReplies, user);
 	generateVisibilityQuery(query, user);
-	if (user) generateMutedUserQuery(query, user);
-	if (user) generateMutedNoteQuery(query, user);
-	if (user) generateBlockedUserQuery(query, user);
-	if (user) generateMutedUserRenotesQueryForNotes(query, user);
+	if (user != null) {
+		generateMutedUserQuery(query, user);
+		generateMutedNoteQuery(query, user);
+		generateBlockedUserQuery(query, user);
+		generateMutedUserRenotesQueryForNotes(query, user);
+		generateMutedUserRepliesQueryForNotes(query, user);
+	}
 
 	if (ps.withFiles) {
 		query.andWhere("note.fileIds != '{}'");