+++ packages/backend/src/server/api/endpoints/notes/timeline.ts
@@ -10,6 +10,7 @@ import { generateMutedNoteQuery } from "@/server/api/common/generate-muted-note-
 import { generateChannelQuery } from "@/server/api/common/generate-channel-query.js";
 import { generateBlockedUserQuery } from "@/server/api/common/generate-block-query.js";
 import { generateMutedUserRenotesQueryForNotes } from "@/server/api/common/generated-muted-renote-query.js";
+import { generateMutedUserRepliesQueryForNotes } from "@/server/api/common/generated-muted-reply-query.js";
 import { ApiError } from "@/server/api/error.js";
 
 export const meta = {
@@ -115,6 +116,7 @@ export default define(meta, paramDef, async (ps, user) => {
 	generateMutedNoteQuery(query, user);
 	generateBlockedUserQuery(query, user);
 	generateMutedUserRenotesQueryForNotes(query, user);
+	generateMutedUserRepliesQueryForNotes(query, user);
 
 	if (ps.includeMyRenotes === false) {
 		query.andWhere(