+++ packages/backend/src/server/api/stream/channels/hybrid-timeline.ts
@@ -76,6 +76,7 @@ export default class extends Channel {
 		if (isUserRelated(note, this.blocking)) return;
 
 		if (note.renote && !note.text && this.renoteMuting.has(note.userId)) return;
+		if (note.replyId != null && this.replyMuting.has(note.userId)) return;
 
 		// 流れてきたNoteがミュートすべきNoteだったら無視する
 		// TODO: 将来的には、単にMutedNoteテーブルにレコードがあるかどうかで判定したい(以下の理由により難しそうではある)