+++ packages/firefish-js/src/api.types.ts
@@ -612,6 +612,9 @@ export type Endpoints = {
 	"renote-mute/create": { req: TODO; res: TODO };
 	"renote-mute/delete": { req: { userId: User["id"] }; res: null };
 	"renote-mute/list": { req: TODO; res: TODO };
+	"reply-mute/create": { req: TODO; res: TODO };
+	"reply-mute/delete": { req: { userId: User["id"] }; res: null };
+	"reply-mute/list": { req: TODO; res: TODO };
 
 	// my
 	"my/apps": { req: TODO; res: TODO };